fix(events): RSVP broken + missing event columns + migration fixes
CRITICAL: event_rsvps table had two bugs:
1. Missing updated_at column — EventService::rsvp() uses ON DUPLICATE KEY
   UPDATE updated_at=NOW() which crashed with "Unknown column". ALL RSVPs
   were broken (422 on every attempt).
2. Status enum too narrow — DB had (going,maybe,declined) but code uses
   (going,interested,not_going,declined). MariaDB silently discarded
   'interested' and 'not_going' values — API returned 200 but data lost.

Also: events table missing is_online, online_link, image_url, video_url
columns (EventService::create referenced them causing 500).

Fixed blocking migration issues:
- federation_topics: added idempotency guards (IF NOT EXISTS)
- listing_reports/images: FK type mismatch (unsignedBigInteger vs int)

+++ database/migrations/2026_04_03_100001_create_listing_reports_table.php
@@ -22,14 +22,14 @@ return new class extends Migration
 
         Schema::create('listing_reports', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('tenant_id');
-            $table->unsignedBigInteger('listing_id');
-            $table->unsignedBigInteger('reporter_id');
+            $table->integer('tenant_id');
+            $table->integer('listing_id');
+            $table->integer('reporter_id');
             $table->enum('reason', ['inappropriate', 'safety_concern', 'misleading', 'spam', 'not_timebank_service', 'other']);
             $table->text('details')->nullable();
             $table->enum('status', ['pending', 'reviewed', 'dismissed', 'action_taken'])->default('pending');
             $table->text('admin_notes')->nullable();
-            $table->unsignedBigInteger('reviewed_by')->nullable();
+            $table->integer('reviewed_by')->nullable();
             $table->timestamp('reviewed_at')->nullable();
             $table->timestamps();
